Add 32 bit casts to bitboards.js

+++ src/bitboard/bitBoard.js
@@ -24,8 +24,8 @@ export default class BitBoard {
     }
 
     and(other) {
-        this.lower = this.lower & other.lower;
-        this.upper = this.upper & other.upper;
+        this.lower = u32(this.lower & other.lower);
+        this.upper = u32(this.upper & other.upper);
         return this;
     }
 
@@ -57,7 +57,7 @@ export default class BitBoard {
             this.lower = this.upper >>> (i-32);
             this.upper = 0;
         } else if (i > 0) {
-            this.lower = (this.lower >>> i) | (this.upper << (32 - i));
+            this.lower = u32((this.lower >>> i) | (this.upper << (32 - i)));
             this.upper = this.upper >>> i;
         }
         return this;
